Latest push

+++ bin/TemplateUpdate.sh
@@ -18,7 +18,12 @@ EOF
 
 ################################################################################
 GB_GITURL="$(git config --get remote.origin.url)"; export GB_GITURL
-GB_GITREPO="$(basename -s .git ${GB_GITURL})"; export GB_GITREPO
+if [ "${GB_GITURL}" == "" ]
+then
+	GB_GITREPO=""; export GB_GITREPO
+else
+	GB_GITREPO="$(basename -s .git ${GB_GITURL})"; export GB_GITREPO
+fi
 
 if [ "${GB_GITREPO}" == "docker-template" ]
 then
@@ -45,5 +50,13 @@ ${DIR}/github-release download \
 	${VERSION} \
 	--name docker-template.tgz
 
-rm -f docker-template.tgz
+if [ -f docker-template.tgz ]
+then
+	echo "# Gearbox[docker-template]: Extracting."
+	tar zxf docker-template.tgz
+	rm -f docker-template.tgz
+	echo "# Gearbox[docker-template]: Done."
+else
+	echo "# Gearbox[docker-template]: Cannot find docker-template repository."
+fi
